Move MPIRandomState to mpirng for reuse.
Also change the interface to accept localsize instead of internally
deciding a localsize. This change is needed for the mockmaker.
